New: Add a history section with original documents
Perhaps it's a bit cutesy, but I also thought this might be an opportunity to save the files as an OCFL object as well. I know OCFL-in-git is a bit redundant, but in the interest of practicing what we preach, this might be a good, quick example that we can point to.

Thoughts and opinions welcome in the PR.