clarify what metadata and content means (#667)
* clarify what metadata and content means

We’ve found that institutions often assume OCFL is opinionated about the internal structure of an object, despite the introduction stating explicitly that OCFL does not define an object’s structure (i.e., what makes an object an object).

One reason for this misunderstanding may be our loose use of the terms’ metadata’ and ‘content’ in the introduction (I’ll take the blame for that). In practice, when people see “metadata,” they often think only of descriptive metadata. However, metadata can also refer to other types, such as METS, which encompasses both structural and administrative metadata, or MIX, which captures technical metadata about an image.

The changes here aim to clarify that we mean all forms of metadata, not just descriptive, so that users better understand OCFL’s scope.

* line length changes

oops......